+++ NetFabric.Hyperlinq.UnitTests/SingleOrDefault.Tests.cs
@@ -82,8 +82,8 @@ namespace NetFabric.Hyperlinq.UnitTests
             result.Should().Be(expected);
         }
 
-        public static TheoryData<int[], Func<int, int, bool>, int> SingleOrDefaultPredicateData =>
-            new TheoryData<int[], Func<int, int, bool>, int>
+        public static TheoryData<int[], Func<int, long, bool>, int> SingleOrDefaultPredicateData =>
+            new TheoryData<int[], Func<int, long, bool>, int>
             {
                 { new int[] { }, (_, __) => true, 0 },
                 { new int[] { 1 }, (_, __) => true, 1 },
@@ -93,7 +93,7 @@ namespace NetFabric.Hyperlinq.UnitTests
 
         [Theory]
         [MemberData(nameof(SingleOrDefaultPredicateData))]
-        public void SingleOrDefaultPredicate_With_ValidIEnumerable_Should_Succeed(IEnumerable<int> source, Func<int, int, bool> predicate, int expected)
+        public void SingleOrDefaultPredicate_With_ValidIEnumerable_Should_Succeed(IEnumerable<int> source, Func<int, long, bool> predicate, int expected)
         {
             // Arrange
 
@@ -106,7 +106,7 @@ namespace NetFabric.Hyperlinq.UnitTests
 
         [Theory]
         [MemberData(nameof(SingleOrDefaultPredicateData))]
-        public void SingleOrDefaultPredicate_With_ValidIReadOnlyCollection_Should_Succeed(IReadOnlyCollection<int> source, Func<int, int, bool> predicate, int expected)
+        public void SingleOrDefaultPredicate_With_ValidIReadOnlyCollection_Should_Succeed(IReadOnlyCollection<int> source, Func<int, long, bool> predicate, int expected)
         {
             // Arrange
 
@@ -119,7 +119,7 @@ namespace NetFabric.Hyperlinq.UnitTests
 
         [Theory]
         [MemberData(nameof(SingleOrDefaultPredicateData))]
-        public void SingleOrDefaultPredicate_With_ValidIReadOnlyList_Should_Succeed(IReadOnlyList<int> source, Func<int, int, bool> predicate, int expected)
+        public void SingleOrDefaultPredicate_With_ValidIReadOnlyList_Should_Succeed(IReadOnlyList<int> source, Func<int, long, bool> predicate, int expected)
         {
             // Arrange
